rename TS interface files
