+++ src/components/ErrorBoundry/interfaces.ts
