--- src/components/Header/IHeaderProps.ts
@@ -1,3 +1,0 @@
-export interface IHeaderProps {
-  children?: React.ReactNode;
-}
