+++ src/components/MovieCard/interfaces.ts
