+++ src/components/MovieList/interfaces.ts
